+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Module change history
 
+## 1.0.2 (March 21, 2019)
+- Added component social.links
+- Added social links settings to module options
+- Modified installer: installs components and admin scripts
+- Success install & uninstall message fix
+
 ## 1.0.1 (March 9, 2019)
 
 - Added migrations instrument (beta).